+++ src/types.ts
@@ -5,7 +5,6 @@ export interface CelatoneContractAddress {
   example: ContractAddr;
 }
 export interface CelatoneConstants extends AppConstants {
-  lcdEndpoint: Record<string, string>;
   maxFileSize: number;
   msgTypeUrl: { [key in MsgType]: string };
 }